pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "macchine"
version = "0.1.0"
description = "Bauer foundation equipment fleet data analysis"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2",
    "pyarrow>=15.0",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "ruff>=0.9.0",
]

[project.scripts]
macchine = "macchine.cli:cli"

[project.urls]
Repository = "https://github.com/vigji/macchine"

[tool.setuptools.packages.find]
include = ["macchine*"]

# =============================================================================
# RUFF - Linting & Formatting
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
]

[tool.ruff.lint.isort]
known-first-party = ["macchine"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]
"tests/*.py" = ["E402"]
"**/__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# PYTEST
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
